Publish protocol as a separate NPM package that can be re-used #1876
