+++ src/protocol/README.md
@@ -1,0 +1,14 @@
+### Replay's Devtools Client
+
+We believe being able to replay your software opens a new world of opportunities. And that our UI should be one of many and probably the most boring one :)
+
+We're committed to helping as much as possible
+
+1. Building a community + ecosystem (discord and friends)
+2. Publishing documentation and examples
+3. Publishing packages that streamline the process
+4. Helping folks to hack on the devtools frontend so they can quickly prototype ideas
+
+**Devtools Client**
+
+Our DevTools frontend does a lot of things, such as showing a video of the recording, adding print statements, etc... This package is our attempt to share that functionality so that it is easy to get started.
